this is the database i worked with
it has three tables. 
The first is users -> id, username, password
the second is inventory -> name, price, image_url
the third is cartitems -> id, username, item, size, color (during the display, it just checks this table and every time the current session's username pops up it keeps a note of that item. This how this table works.